// shared/bsdd-api/generated/types.gen.ts
export type ClassListItemContractV1Classes = {
    uri?: string | null;
    code?: string | null;
    name?: string | null;
    classType?: string | null;
    referenceCode?: string | null;
    parentClassCode?: string | null;
    /**
     * Description of the class, max 101 characters long
     */
    descriptionPart?: string | null;
    children?: Array<ClassListItemContractV1Classes> | null;
};
// ...
export type CodeNameDto = {
    code?: string | null;
    name?: string | null;
};
// ...
export type DictionaryContractV1 = {
    code: string;
    uri: string;
    name: string;
    version: string;
    organizationCodeOwner: string;
    organizationNameOwner: string;
    changeRequestEmail?: string | null;
    /**
     * The default language for this domain
     */
    defaultLanguageCode: string;
    isLatestVersion: boolean;
    isVerified: boolean;
    isPrivate: boolean;
    /**
     * Name or short description of the license under which you can use this data
     */
    license?: string | null;
    /**
     * URL where you can find more details about the license or default link to info about licensing
     */
    licenseUrl?: string | null;
    /**
     * Name or short description of the quality assurance procedure used while creating and maintaining the domain data
     */
    qualityAssuranceProcedure?: string | null;
    /**
     * URL where you can find more details about the quality assurance procedure
     */
    qualityAssuranceProcedureUrl?: string | null;
    /**
     * For state indication, like "Preview", "Active", "InActive", "Released
     */
    status?: string | null;
    /**
     * Url to site with more info on domain
     */
    moreInfoUrl?: string | null;
    /**
     * Date of the release of this version
     */
    releaseDate?: string | null;
    /**
     * Date and time the data of the domain has been last updated in bSDD
     */
    lastUpdatedUtc?: string;
    /**
     * List of (partly) available languages for this dictionary
     */
    availableLanguages?: Array<CodeNameDto> | null;
};